+++ src/pages/rooms-overview.vue
@@ -164,12 +164,7 @@ export default {
 			if (JSON.stringify(this.draggedElement.from) == JSON.stringify(pos))
 				return;
 
-			if (
-				(this.draggedElementName == "vRoomAvatar" ||
-					"vRoomGroupAvatar" ||
-					"groupItem") &&
-				toElementName == "vRoomEmptyAvatar"
-			) {
+			if (toElementName == "vRoomEmptyAvatar") {
 				this.savePosition();
 			}
 			this.showDeleteSection = false;
@@ -182,7 +177,8 @@ export default {
 				return;
 
 			if (
-				(this.draggedElementName == "vRoomAvatar" || "groupItem") &&
+				(this.draggedElementName == "vRoomAvatar" ||
+					this.draggedElementName == "groupItem") &&
 				toElementName == "vRoomAvatar"
 			) {
 				this.savePosition();
@@ -196,7 +192,8 @@ export default {
 				return;
 
 			if (
-				(this.draggedElementName == "vRoomAvatar" || "groupItem") &&
+				(this.draggedElementName == "vRoomAvatar" ||
+					this.draggedElementName == "groupItem") &&
 				toElementName == "vRoomGroupAvatar"
 			) {
 				this.savePosition();